fix: env variables (#36)
## fixes KILTProtocol/ticket#3076
The environment variables for OpenDID were only utilized during the
build process, resulting in the use of default values specified in the
code. While OpenDID should be able to run on Spiritnet, the environment
variables need to be configurable.

To address this issue, Parcel is replaced with react-scripts. This
change is made because Parcel lacks real configuration tools to adjust
bundled files. The config.js file in the public folder was bundled into
a different file, making it challenging to substitute.

With the new approach, the environment variables are stored in a
config.js file. It is now possible to configure them using different
orchestration software such as k8s or Docker Compose. A different
configuration file can now be mounted into the dist folder.